requirements.txt
asgiref==3.8.1
Brotli==1.1.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==44.0.3
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
ecdsa==0.19.1
gunicorn==23.0.0
idna==3.10
packaging==24.2
pyasn1==0.4.8
pycparser==2.22
python-jose==3.4.0
requests==2.32.3
rsa==4.9.1
six==1.17.0
sqlparse==0.5.3
urllib3==2.4.0
whitenoise==6.9.0
